Reorganized directory structure, scripts --> stages
