+++ stages/run_opensfm.py
