+++ stages/splitmerge.py
